chore: Linux (gcc) でビルド可能にするクロスコンパイル修正
AppleClang では通るが libstdc++ / gcc では落ちる問題を修正。macOS ビルド
の挙動は変わらない (mmw-render / MikuMikuWorld.app を再ビルドして
--dry-run で回帰なしを確認済み)。

- Platform/Stubs/sal.h: __in / __out / __inout スタブを削除。libstdc++ の
  <bits/locale_conv.h> が内部変数として __out を使っているため、マクロ化
  すると無関係な翻訳単位で parse エラーになる (DirectXMath 自身は古い SAL
  アノテーションを参照していない)
- Rendering/Camera.h: XMVECTOR の brace-init を XMVectorSet に置き換え
  (aarch64 NEON では float32x4_t の戻り値 brace-init が通らない)
- Audio/Sound.{h,cpp}: ma_sound_get_* / ma_dr_*_open_memory の out パラ
  メータを ma_uint64 に合わせる (Linux では uint64_t と ma_uint64 が別型)
- UI.h: UI::addSelectProperty<FlickType> の explicit specialization を
  class 内から namespace スコープに移動 (gcc は非 namespace scope の
  explicit specialization を拒否)
- Particle.cpp / ScoreEditorTimeline.cpp: std::ceilf -> std::ceil
- MathUtils.h / IO.cpp: 明示的に <cmath> / <cstring> をインクルード
- File.cpp / File.h: getStreamMode の戻り値を std::ios_base::openmode に
- 各ヘッダに <cstdint> を追加 (AppleClang は他のヘッダ経由で引っ張って
  いたが libstdc++ では引っ張らない)
- CMake: Linux では _XM_NO_INTRINSICS_ を定義して DirectXMath のスカラ
  実装にフォールバック、X11::X11 をリンク
- .docker/Dockerfile.linux-render: Ubuntu 24.04 + Mesa llvmpipe + Xvfb +
  ffmpeg で mmw-render をヘッドレスに実行できるイメージ

+++ MikuMikuWorld/Particle.cpp
@@ -81,7 +81,7 @@ namespace MikuMikuWorld::Effect
 		{
 			if (burst.cycles == 0)
 			{
-				float cycleCount = std::ceilf((ref.duration - burst.time) / burst.interval) + 1;
+				float cycleCount = std::ceil((ref.duration - burst.time) / burst.interval) + 1;
 				count += cycleCount * burst.count;
 			}
 			else
